requirements.txt
ccxt
tweepy
pprint
time
threading
python-dotenv
datetime
numpy
